move ledger services implementation to own sub-module
